requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
pydantic==2.9.2
pydantic-settings==2.6.0
email-validator==2.2.0
python-dotenv==1.0.1
anthropic==0.40.0
openai==1.54.0
httpx==0.27.2
anyio==4.6.2
typing_extensions==4.12.2
annotated-types==0.7.0
streamlit==1.41.0
pandas==2.2.0
plotly==5.18.0
